# backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: SmartPresence API
  description: API for managing users, rooms, classes, sessions, and attendance in an educational institution
  version: 1.0.0
  contact:
    name: SmartPresence Development Team
    email: dev@smartpresence.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.smartpresence.com/api
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /mobile/students/login:
    post:
      tags:
        - Mobile
      summary: Student mobile login
      description: Authenticate student via matric number and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [matricNo, password]
              properties:
                matricNo:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
        '401':
          description: Invalid credentials
  /mobile/me:
    get:
      tags:
        - Mobile
      summary: Get my profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Student profile
  /mobile/me/classes:
    get:
      tags:
        - Mobile
      summary: Get my classes
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Classes list
  /mobile/me/sessions:
    get:
      tags:
        - Mobile
      summary: Get my sessions
      parameters:
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Sessions list
  /mobile/me/attendance:
    get:
      tags:
        - Mobile
      summary: Get my attendance
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Attendance list
  /mobile/attendance/mark:
    post:
      tags:
        - Mobile
      summary: Mark attendance
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarkAttendanceRequest'
      responses:
        '201':
          description: Attendance marked
        '403':
          description: Forbidden (validation failed)
        '409':
          description: Device already used in session
  /mobile/me/notifications:
    get:
      tags:
        - Mobile
      summary: Get student notifications
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: unread_only
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Notifications list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationsResponse'
  /mobile/me/notifications/unread-count:
    get:
      tags:
        - Mobile
      summary: Get unread notification count
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Unread count
          content:
            application/json:
              schema:
                type: object
                properties:
                  unread_count:
                    type: integer
  /mobile/notifications/{id}/read:
    put:
      tags:
        - Mobile
      summary: Mark notification as read
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Notification marked as read
          content:
            application/json:
              schema:
                type: object
                properties:
                  notification_id:
                    type: integer
                  is_read:
                    type: boolean
                  read_at:
                    type: string
                    format: date-time
        '404':
          description: Notification not found
  /mobile/notifications/mark-all-read:
    put:
      tags:
        - Mobile
      summary: Mark all notifications as read
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All notifications marked as read
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  updated_count:
                    type: integer
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate a user and receive a JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  minLength: 6
                  example: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BaseResponse:
      type: object
      required:
        - success
        - message
        - timestamp
      properties:
        success:
          type: boolean
        message:
          type: string
        timestamp:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - timestamp
        - path
        - method
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - message
            - type
            - statusCode
          properties:
            message:
              type: string
            type:
              type: string
            statusCode:
              type: integer
            details:
              type: array
              items:
                type: object
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        method:
          type: string
        requestId:
          type: string

    # User Schemas
...
    LoginResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              required:
                - token
                - user
              properties:
                token:
                  type: string
                user:
                  type: object
                  required:
                    - id
                    - email
                    - role
                  properties:
                    id:
                      type: integer
                    email:
                      type: string
                      format: email
                    role:
                      type: string
                      enum: [admin, teacher, student]

    # Room Schemas
...
    MarkAttendanceRequest:
      type: object
      required:
        - class_id
        - session_id
        - wifi_ssid
        - device_id
      properties:
        class_id:
          type: integer
          minimum: 1
        session_id:
          type: integer
          minimum: 1
        wifi_ssid:
          type: string
          minLength: 1
          maxLength: 100
        bluetooth_beacon_id:
          type: string
          maxLength: 100
        device_id:
          type: string
          minLength: 3
          maxLength: 200
...
    Notification:
      type: object
      properties:
        notification_id:
          type: integer
        type:
          type: string
          enum:
            [
              session_reminder,
              attendance_confirmed,
              session_cancelled,
              class_enrolled,
            ]
        title:
          type: string
        message:
          type: string
        is_read:
          type: boolean
        related_session_id:
          type: integer
        related_class_id:
          type: integer
        created_at:
          type: string
          format: date-time
        read_at:
          type: string
          format: date-time
        class_name:
          type: string
        session_start_time:
          type: string
          format: date-time
    NotificationsResponse:
      type: object
      properties:
        notifications:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer
            hasNext:
              type: boolean
            hasPrev:
              type: boolean
...